move functions around
