+++ synse_server/cmd/info.py
@@ -1,4 +1,9 @@
 
 
 async def info():
-    """"""
+    """Generate the device info response data.
+
+    Returns:
+        dict: A dictionary representation of the device info response.
+    """
+    pass
