+++ synse_server/cmd/plugin.py
@@ -1,12 +1,43 @@
 
+import synse_server.plugin
 
-async def plugin():
-    """"""
+
+async def plugin(plugin_id):
+    """Generate the plugin response data.
+
+    Args:
+        plugin_id (str): The ID of the plugin to get information for.
+
+    Returns:
+        dict: A dictionary representation of the plugin response.
+    """
+    pass
 
 
 async def plugins():
-    """"""
+    """Generate the plugin summary response data.
+
+    Returns:
+        list[dict]: A list of dictionary representations of the plugin
+        summary response(s).
+    """
+    summaries = []
+    for p in synse_server.plugin.manager:
+        summaries.append({
+            'active': True,
+            'description': p.description,
+            'id': p.id,
+            'maintainer': p.maintainer,
+            'name': p.name,
+        })
+
+    return summaries
 
 
 async def plugin_health():
-    """"""
+    """Generate the plugin health response data.
+
+    Returns:
+         dict: A dictionary representation of the plugin health.
+    """
+    pass
