+++ synse_server/cmd/read.py
@@ -10,3 +10,13 @@ async def read_device():
 
 async def read_cache():
     """"""
+
+
+def readings():
+    """Generate the readings response data.
+
+    Returns:
+        list[dict]: A list of dictionary representations of device reading
+        response(s).
+    """
+    pass
