+++ synse_server/cmd/scan.py
@@ -1,4 +1,10 @@
 
 
 async def scan():
-    """"""
+    """Generate the scan response data.
+
+    Returns:
+         list[dict]: A list of dictionary representations of device
+         summary response(s).
+    """
+    pass
