+++ synse_server/cmd/tags.py
@@ -1,4 +1,9 @@
 
 
 async def tags():
-    """"""
+    """Generate the tags response data.
+
+    Returns:
+        list[string]: A list of all tags currently associated with devices.
+    """
+    pass
