+++ synse_server/cmd/test.py
@@ -1,5 +1,14 @@
 
+from synse_server import utils
 
-# FIXME (etd): probably dont need this..
+
 async def test():
-    """"""
+    """Generate the test response data.
+
+    Returns:
+        dict: A dictionary representation of the test response.
+    """
+    return {
+        'status': 'ok',
+        'timestamp': utils.rfc3339now(),
+    }
