+++ synse_server/cmd/transaction.py
@@ -6,3 +6,12 @@ async def transaction():
 
 async def transactions():
     """"""
+
+
+def txn():
+    """Generate the transaction response data.
+
+    Returns:
+        dict: A dictionary representation of the transaction response.
+    """
+    pass
