+++ synse_server/cmd/version.py
@@ -1,5 +1,14 @@
 
+import synse_server
 
-# FIXME (etd): probably don't need this
-async def version():
-    """"""
+
+def version():
+    """Generate the version response data.
+
+    Returns:
+        dict: A dictionary representation of the version response.
+    """
+    return {
+        'version': synse_server.__version__,
+        'api_version': synse_server.__api_version__,
+    }
