+++ synse_server/cmd/write.py
@@ -6,3 +6,13 @@ async def write_async():
 
 async def write_sync():
     """"""
+
+
+def write():
+    """Generate the asynchronous write response data.
+
+    Returns:
+         list[dict]: A list of dictionary representations of asynchronous
+         write response(s).
+    """
+    pass
